Setup for translation
To start translation, we need to initially import the translation
file - and place it at the proper place so that the usual CI
scripts can handle it. The proper place is for all python
projects $PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports will be done by the OpenStack Proposal bot.

Change-Id: I4223ec5ffd3770ef90e23cd8f84341097c1fe84f