Remove ICMPV6_ALLOWED_TYPES
We can do this now quickly because it is not used outside of the
neutron core tree yet:
http://codesearch.openstack.org/?q=ICMPV6_ALLOWED_TYPES

Why? Because ICMPV6_ALLOWED_TYPES is not a "constant" -- it is more
like a policy. It turns we need to change the list of allowed ICMPv6
types. This is being taken care of by
https://review.openstack.org/310954

Change-Id: I81f5546251efe6cef1f79a7889cd9b15369d7f79