Add subport validator for vlan-aware-vms
This patch introduces a new validator method for a list of trunk
subports.

A subport input is defined  by a ternary tuple of 'port_id' and
optionally 'segmentation_id' and 'segmentation_type'. If the
segmentation information is specified, both type and id must be
present; port_id must be a valid UUID and unique within the list,
and the same goes for segmentation id. More implementation specific
validations are deferred to the plugin layer.

The validator may find use in other projects besides 'neutron' alone
and thus is contributed in neutron-lib sooner rather than later.
Alternatively local validators should be supported to avoid modifying
the module variable as reported in bug 1584237.

Partially-implements: blueprint vlan-aware-vms

Change-Id: Ic04c547a75a7851d268ec6121eceaf0be9ca3586