Update validator accessors
In preparation for making access to the validators structure
private, this patch adds a getter to access the validators,
support for 'type:' validation_type prefixing and a comparison
of validator function source to determine validation registration
collisions.

Comparing the source of validators is used to allow consumers
to load modules that reinitializes the validator registration
call site. This is done, for example, via [1] which is used during
neutron unit tests as well as loading extensions. Note that
other alternatives are viable, but this approach puts the
least amount of burden on consumers while still remaining
sane.

[1] https://docs.python.org/2/library/imp.html#imp.load_source

Change-Id: I19ea55bcfd7944e202d83864b555dc45f770a3f4
Partial-Bug: #1584237