Localized exception message hacking check
As per [1], this patch adds a new hacking check that verifies
raised exceptions don't use unlocalized string messages. This
patch also adds unit tests for the new check and enables
checks for neutron-lib itself (as per tox.ini).

Note: Perhaps this check better belongs in openstack-dev/hacking
but it seems we already have a number of more generic checks
in neutron. We can discuss that here need be.

NB: Without [1] this new hacking check fails when running pep8 on
neutron-lib.

[1] https://review.openstack.org/#/c/324374/

Change-Id: Ia13f190ec6843ab360d115871d345c4ddc302cc5