WADL to RST migration
This patch brings the Networking API Reference to a repository
under the neutron stadium. It is a part of the effort described here [1]:
As we discussed in the stadium evolution [2], we plan to have the API stuff
in neutron-lib, so the API reference is hosted on neutron-lib as well.

The goal of this patch is to import converted RST API references.
Cleanup is planned once this patch is merged.

[1] https://wiki.openstack.org/wiki/Documentation/Migrate#API_Reference_Plan
[2] https://review.openstack.org/312199

Change-Id: I6f2c4d476c31b252bfb181138505673f397639fc