Generate API report tooling
NB: The code here is not pretty and could use some love.
However there's many more pressing things to do, so perhaps
this can be revisited later need be (as we agreed during the
neutron-lib meeting [3]).

This patch adds a python and bash script to our tools that can be
run during tox/gate to generate python public API reports as
discussed during out neutron lib meeting [2][3]. The idea is that
this tool can be run to generate a public API report that we can
use to release note our neutron-lib releases [3].

There are 2 main approaches for extracting an API from python
source; (a) loading the source an using inspect based modules
to reflect the API and (b) reading the source without loading
it and parsing API signatures from raw source text. In addition
a hybrid approach can be taken with uses a combination of

This script uses preprocessing and inspection to permit python
source to be loaded (without its dependencies) and inspected.
The inspected source's public signatures are collected and
stored in a JSON format that defines the public API for the
inspected source. JSON API reports can also be compared to
determine what public APIs have changed, added, removed or
stayed the same.

The high level approach for generating an API is as follows:
- Copy the source directory tree to a temp staging location.
- Replace all imports in the staging source with mock imports
so that modules can be loaded without dependencies.
- Walk the source, dynamically loading modules and inspecting
them.
- Build a representation of the inspection including public
signatures. This include public functions, methods, and
variables. Only simple/scalar variable (module or class attrs)
values are tracked at present.
- Format the representation into a JSON format that can
be saved to a file and later used for comparison.

For a sample of the current output, see [1].

This patch also adds a standalone tox target called 'api-report'
that can be used to generate a new API report against the
current change. Note that tox target explicitly specifies the
neutron-lib version to do the api diff against and thus will need
updating per neutron-lib release.

[1] http://paste.openstack.org/show/526626/
[2] http://goo.gl/9Xpi4K
[3] http://goo.gl/LlAjX7

Change-Id: If31ae832f86ff146bd8fb53b6d4a5aa9b8217f72