devref for public API docstring
This patch updates our review guidelines with respect to
public API docstrings. While we want to encourage docstrings,
we don't want hard enforcement that may result in a raised
entry barrier for lib contributors.

I'll investigate the ability to implement a warn-only hacking check
for docstrings. This would allow us to identify and cleanup
pydoc "debt" if we desired.

Change-Id: I73734e4acc34ed7716e3daceb664cebbf045c8ff