Sync neutron callbacks into lib
Update neutron-lib callbacks with the latest from neutron
in preparation for updates to the callback notification interface.
The callback updates also require a few module level functions
in the db package.

A few additional unit tests are also added for coverage as is
python docstrings for the public API contained herein.

Change-Id: I74effcc4a564a92c418132042bbdd5bed22c4950