Enhance pyir tooling CLI
This change updates the pyir report tooling:
- Regex patterns for blacklisting report generation and
diff are now supported via CLI.
- CLI commands are modularized for easier extension.
- A new CLI command is added that takes an existing JSON
API report and outputs its signatures to STDOUT. This is
useful for listing an API given a report.
- A few minor bugs are fixed that were found during testing.

Note:
This is the last major "enhancement" I have planned for this
tooling code in its current form. Of course bugs will be
addressed. Once we give it a little burn-in time (assuming we
find it useful) I'll query the broader community for interest.
If there's broader interest I'll consider refactoring + hosting
it as a separate project (obviously a ways out, as that takes
time).

Change-Id: Ia6ed70bd9457e119a6c12d8d133f77384e392282