Add converter to convert IPv6 addresses to canonical format
With respect to Section-4 of [1], the following patch adds
a function to compress the extended IPv6 address passed to
the neutron server to its canonical form. This allows
compressed IPv6 addresses to be stored and remove any
leading zeroes in the IPv6 addresses.

Non-canonical IPv6 address formatted inputs wont be blocked,
but the addresses would be normalized to canonical formats.

[1]: http://tools.ietf.org/html/rfc5952

Partial-Bug: #1531103
Change-Id: Ic9afa5f90fba783748f7c56e8fba02e06af69748