api-ref: introduced a stub sub-page to cover general info
Apart from specifics of features like L2 or L3 or QoS API, there should
be a separate place to contain general information about how to use
Neutron API. This patch introduces an empty stub sub-page for just that.
Follow-up patches will expand with details.

Things like sorting and pagination behaviour, or authentication, or
headers, are good candidates for this sub-page, because they are in most
cases not resource specific but cover the whole Neutron API. It's better
to keep those details in a single place instead of duplicating the
information for each new resource.

Change-Id: Ib04b18bf9e872fd2471a7f3dda3853349063761a