Move valuable API info from neutron-specs repo
This information is stored (unlinked) in neutron-specs repo under
misc/api/networking_general_api_information.rst This patch just moves
the contents into api-ref, where they seem to belong.

Note that the contents are not current in some parts. For example, it is
said there XML format of requests is still supported, which is not true
for quite a while. This contents will be cleaned up in follow up
patches. For now, I keep the text intact to ease reviewer's life.

Adding two documentation bugs to be closed by the patch because the
information missing in api-ref is already included in the bits copied by
this patch.

Closes-Bug: #1609934
Closes-Bug: #1609932
Change-Id: Ic0c2348d7db941deaec6e1d110954fe5893c2265