Enhance valid_values to use __contains__
Actually valid_values doesn't cover many of the use cases or validates
that the value and valid_values provided can be used with 'in' call.

his patch tries to improve this situation to ensure that both are valid
for being checked or fail otherwise plus validating that valid_values
contains the value provided.

Closes-Bug: 1604073
Change-Id: I47a6f7d58d6cbeb8fd245620d187c4c565497290