Migrate neutron agent extensions to neutron-lib
Migrate neutron agent extensions to neutron-lib.  This includes both the
common abstract object as well as the l2 and l3-specific objects for
agent extensions.

This functionality is used by neutron, neutron-fwaas, networking-sfc,
and tap-as-a-service.

Please note that this does not migrate the l3_agent_extensions_api to
neutron-lib.  This is because that specifically has a long dependency
chain that ends up requiring agent/[linux,windows]/utils.py, which
should be migrated separately.

Change-Id: I7255a5ae6b706c09acecc0dea5dbd2febae0c282
Needed-By: I6d2cff03207924c7492c6136d4b21a9817769c67
Needed-By: If9235c1cf39b8ad2ef049e8154b82736e1f1d763
Needed-By: Ie8f0e70b11435d627f73e3fed45f946e863e90bc
Needed-By: Ia5962acfd737b4b350b94591802928a5a4dc53fb