Introduce Plugin Directory for Neutron
The NeutronManager is heavily relied on across Neutron projects
primarily to look up plugins at runtime. During the init phase
however, the Manager lays down all sorts of plumbing needed for
other core objectives. Rehoming it as is would be not only
impossible, but also perpetrating the bad pattern where the
class itself holds too many responsibilities.

This patch extract the part whose responsibility is the mapping
between aliases and plugins. The NeutronManager will then make
use of this registry to store the plugin mapping, and subprojects
will stop using the manager in favor of the newly introduced
plugin directory.

Partially-implement: blueprint neutron-lib

Change-Id: I83b0217ea16d2c3f4ada05686c50470e50bd5208