api-ref: project_id in req/resp body should be "body"
project_id and tenant_id field in request/response body are
marked as "path" now. It should be "body".

- "project_id-path" already exists and it can be used
  for "project_id" in URL path.
- "project_id" is now used for body.
- "project_id-body" is now duplicated, so it was removed.
  fwaas.inc is the only user of project_id-body and
   it is updated accordingly.
- quotas.inc is updated to use 'project_id-path'.
  Also project_id and tenant_id in response body of a quotas operation
  have been dropped as they do not exist.

Note that project_id/tenant_id in request body should be marked
as "optional" in most resources but this patch does not touch them
to avoid unnecessary merge conflicts.
They will be fixed in separate patches.

Change-Id: Ic5a4f55b837ee0a51b7186c3342a94c8c00f6c97
Closes-Bug: #1650174