Add dummy API definitions for extension subresources
This change introduces the dummy definitions for extension
subresources, and its related testing coverage. It was
observed in a few definition proposals (e.g. [1]) that this
was a missing element in the description of a neutron extension.

The enhancement got reflected in the existing definitions, such
as trunk and trunk_details, and thus amends were made to these
to show how to use the dummy suggestions in practice.

[1] https://review.openstack.org/#/c/413182

Change-Id: I5eb36cdf99039f2e74c93735fcddd5953c7fe086