Use a single hacking check factory
While the intention of the incubating hacking check
factory was to provide a means to roll-out new checks
without breaking consumers [1], the complications involved
with running 2 factories in a gate job [2] perhaps negate
the benefits of this approach.

Therefore, as per discussion with Armando (see the neutron
IRC logs from Jan 9, 2017 at 18:58) this patch proposed
we only have a single hacking check factory for all consumers.

The implications are as follows:
- All new adopter hacking checks go into the single factory.
- Consumers failing the new pep8 hacking check will
see the failure via neutron-lib periodic job.
- Consumers fix failures ASAP.

IFF a special case arises and we need to release neutron-lib
before consumers adopt check(s), we can comment them
out of the factory to mitigate as a whole. Or individual
consumer project's can 'ignore' failing checks as outlined
in the usage.rst of this patch.

This patch updates the dev-ref as well as the code to
implement the above scheme.

[1] https://review.openstack.org/#/c/350723/
[2] https://review.openstack.org/#/c/412628/

Change-Id: I0a8b39052c0f81f2ecbab3923353be1537b89a65