api-ref: Fix LBaaS v1 API reference
LBaaS v1 support was removed in Newton release, but there are
still users of LBaaS v1 API. It looks reasonable to maintain it
with minimum effort rather than dropping it.
This commit fixes LBaaS v1 API document with rough checks.

Note that project_id field is not listed in LBaaS v1 API intentionally
as project_id field was added in Newton release.

Part of blueprint neutron-in-tree-api-ref

Change-Id: I82d8f00f726247eb305a6b7207ae1bd2262fe296