Formalize base service classes
The PluginInterface and WorkerSupportServiceMixin classes
were rehomed to neutron-lib recently [1] and included a private
version of the respective classes. Based on the existing usages
of such [2][3], this patch:
- Renames WorkerSupportServiceMixin as WorkerBase making it
public for consumption [2].
- Collapses _PluginInterface into ServicePluginBase making the
definition more compact and understandable.
- Formalizes the use of ServicePluginBase by updating the
docstring in ExtensionDescriptor to indicate an instance of
ServicePluginBase is used for plugin interfaces. This is kosher
as existing uses today all return extension plugin interfaces that
are subclasses of ServicePluginBase and not PluginInterface.

Also see related neutron consumption patch [4].

[1] https://review.openstack.org/412519/
[2] http://codesearch.openstack.org/?q=WorkerSupportServiceMixin
[3] http://codesearch.openstack.org/?q=PluginInterface
[4] https://review.openstack.org/#/c/441129

Change-Id: I722252b845a4396f1b193459b22e4352c93241e6