Correctly set project_name
With newer keystone middleware tenant* isn't set, project* is. The
current neutron logic assumes tenant is the good values, and that
project should mirror it, but that logic actually needs to be
inverted.

In order to minimally impact unit tests as operations, the strategy
here is to just prefer project_name, but fall back to tenant_name if
project_name isn't filled out (this should basically never be the
case).

With this project_name shows up in the context log lines in devstack
again (it's been missing for some time).

Original-Commit: Ice6a4d046b492f02e3306627a760e5f4573f0fc6
Co-authored-by: Sean Dague <sean@dague.net>

Change-Id: Ibea8c379328c8afd009df3f37019a6ba0b106d01