rehome extra_dhcp_opt api-def
This patch rehomes the extra_dhcp_opt extension
from neutron.extensions into the API definitions
within neutron-lib. As part of the rehome some
refactoring is done:
- The DHCP opt exceptions are removed; one is never used [1]
and the other extends InvalidInput and is only used in the
validator [2] and can be replaced by a vanilla InvalidInput
exception.
- The validator is refactored and moved into validators.
- The 'id' EXTRA_DHCP_OPT_KEY_SPECS is removed as its
unused.
- The validator in lib is registered as type
'type:list_of_extra_dhcp_options' which is slightly
different than the original type. This is intentional to
ensure the rehome/consume goes smoothly and the
same validator is not registered multiple times.

UTs + release note also included.

In addition a bugfix is included that's required for the validator
herein. The problem is that if data is passed into
_validate_list_of_items that is a list of dicts, a TypeError is
raised because dicts are unhashable.

This api-def is used in stadium sub-projects such as
networking-ovn and is therefore a candidate for
rehoming.

[1] http://codesearch.openstack.org/?q=ExtraDhcpOptNotFound
[2] http://codesearch.openstack.org/?q=ExtraDhcpOptBadData

Change-Id: Ic55b6a1db38634cab7854538f77632115f1b7ec6