rehome plugin common constants
This patch rehomes a bulk of constants from
neutron.plugins.common.constants as they are used
by a number of our sub-projects [1].

Notable differences include:
- Service operation status constants are not defined in the
plugin constants as they are already available in
neutron_lib.constants.
- The DEFAULT_SERVICE_PLUGINS is not included. This one
appears to be neutron specific and should likely be made
private in neutron.
- The EXT_TO_SERVICE_MAPPING dict is not included. We'll
push this in a separate change so we can discuss in greater
detail.

A release note is also included.

[1] http://codesearch.openstack.org/?q=from%20neutron
%5C.plugins%5C.common%20import%20constants

Change-Id: I6e86d737064f97d82e36a4e82a9f2a22fd38febe