Tweak trunk extension to support Ironic use case
To support the Ironic use case, the trunk extension
needs to know about provider details for the network.

This patch documents this relationship, and it adds
support for the 'inherit' segmentation type that
acts as a flag that a user can use to signal his/her
request in having the segmentation details derived
from the underlying network.

The API has not practically changed, and thus no
release note or API extension shim is required for
this patch.

Related-bug: #1648129

Change-Id: Ib510aade1716e6ca92940b85245eda7d0c84a070