Stop enforcing non-debug log translations
This is in line with discussion [1] where translators asked to stop
translating log messages, leaving translations for user visible messages
only.

This patch kills N531 hacking check. A follow up in one of next library
releases will also kill N533, replacing it with a more broad check
forbidding any log translations, not just debug.

We don't squash existing check removal with the patch that adds new
checks because those checks are mutually exclusive, and so with a
squash, we would make it so that consuming projects would not be able to
adopt their code to the coming changes in a non-breaking way.

[1] http://lists.openstack.org/pipermail/openstack-dev/2017-March/114191.html

Change-Id: I5b92dbda79bd45cd1e2f5329ebc046a2dd9b7bff