Add flush_on_subtransaction=True to engine config
By default, the new enginefacade doesn't emulate the same
autoflush behavior we had on exits of session context managers
with the old style. As of olso.db 4.18, we have an option to
get this same behavior. This patch implements it in our db.api
to help with the adoption of the new enginefacade by providing
the same flush semantics.

Related-Bug: #1664643
Partially-Implements: blueprint enginefacade-switch
Change-Id: I71e274d0f0199896f28daae6db71241d9b43a2f3