Patch _get_callback_manager for callback fixture
The existing version of the callback fixture patches the global var
_CALLBACK_MANAGER, which, is returned from the callback registry
via _get_callback_manager() using a singleton checking pattern.
While this is fine in a number of cases, the implementation of
_get_callback_manager() will enclose a reference to the
_CALLBACK_MANAGER under certain testing conditions causing the fixture
to not work properly.

This patch proposes that we just mock the _get_callback_manager() return
value, which works in all cases.

No release note included because this only changes internal semantics.

Change-Id: I9966c90e3f90552b41ed84a68b19f3e540426432