APIDefinitionFixture bugfix
Currently APIDefinitionFixture backs up the contents of its target dict
(attributes.RESOURCES and/or an API def's RESOURCE_ATTRIBUTE_MAP) and
on restore replaces the respective target with its backup copy. While
this works in a number of cases, it doesn't take into account "longer"
references to the respective dict. In some cases consumers hold a
reference to an attribute dict longer than the duration of the fixture
in which case their reference is not cleaned up on restore.

This patch address this bug and includes UTs. No release note is
included as this is only an internal behavior change.

Change-Id: Id61ba4e4649d3a73c9397632755f1cd0c38d7ab3