Withdraw doc for bgpvpn-routes-control API extension
Next release feature freeze is close, and it's not realistic to expect
that the implementation-in-progress of the bgpvpn-routes-control API
extension will be ready in time.

To avoid confusion, this change reverts the parts of change
Ib2cefcb6a345887435b754fbdcff3584329d45d2 that are documenting
the new bgpvpn-routes-control API extension.

The intent is to revert this change once an implementation is finished.

Related: blueprint routes-control

Change-Id: I97146127959cc0161892da5fef65c2a4a1598886