Rehome FAULT_MAP from the neutron API
The fault map enables a developer to map neutron exceptions to
WebOb Exceptions. This enables the neutron service to raise the
relevant exception.

The FAULT_MAP is now maintained in neutron-lib. This dictionary
can be extended if necessary, for example a 3rd party plugin
wants to map plugin exceptions to the relevant WebOb exceptions.

The default mappings are:

    exceptions.NotFound: webob.exc.HTTPNotFound
    exceptions.Conflict: webob.exc.HTTPConflict
    exceptions.InUse: webob.exc.HTTPConflict
    exceptions.BadRequest: webob.exc.HTTPBadRequest
    exceptions.ServiceUnavailable: webob.exc.HTTPServiceUnavailable
    exceptions.NotAuthorized: webob.exc.HTTPForbidden
    netaddr.AddrFormatError: webob.exc.HTTPBadRequest
    policy.PolicyNotAuthorized: webob.exc.HTTPForbidden

Change-Id: Idfd956fa213e1ba6c3b1c983f246c5b515951e6d