Revert "hardware offload support for openvswitch"
This reverts commit f39c1a1615cc64e08aff7e873b62def53638659e.

As it turned out during nova review for the feature [1], we can avoid a
separate vif type, reusing the normal one. This patch drops the type
added in the previous library release.

In theory, it's not backwards compatible, but since there is probably
little adoption for the last library release, and especially for the
new type (I would be surprised to learn anyone started using it), I
think it's ok to drop it and pretend it never happened.

[1] https://review.openstack.org/#/c/398265/

Related-Bug: #1627987

Change-Id: Ic1ee2cd4b0c75b2d0fe5f672e18b9e6206bb7f41