rehome dvr extension api def
This patch rehomes the neutron dvr extension API definition into
neutron-lib. In addition the MacAddressGenerationFailure is defined
twice in neutron; once for net failures and once for host failures.
As both exceptions are used today [1], this patch proposes we name
them slightly different to reflect their usage, but keep them in
the generic exceptions module as it's foreseeable that MAC address
failures can happen outside the context of the dvr extension.

UTs and release note are also included.

[1] http://codesearch.openstack.org/?q=MacAddressGenerationFailure

Change-Id: I6d98e83287833882515240d7db1faa4a1deb7da1