rehome qos api definition
This patch rehomes neutron's qos extension's API definition into
neutron-lib. UTs and a release note are also included.

Reviewers:
Please note the difference in the SUB_RESOURCE_ATTRIBUTE_MAP attributes
herein vs those in the original. Here we've added conversion to int on
a few attributes and also removed the default value of None for those.

Change-Id: I8ab3712850ec02278b3efb9161fc03236926b842