Rehome net-mtu-enh extension definition
This new extension allows to set mtu on create and update. Plugins are
free to implement only create if needed, returning a meaningful error if
update is not supported. Plugins should also make sure that MTU makes
general sense in context of corresponding deployment.

Plugins supporting the new extension are expected to signal support for
net-mtu too because the new extension is a superset of the old feature.

We can't just update the existing net-mtu extension because then support
for writes is not discoverable.

Change-Id: I3dd29578cee4ac4f9f55f6500965664e9e8069a6
Related-Bug: #1671634