Adds IPinIP protocol
Neutron doesn't support the IP-in-IP protocol by name in
security group rules. With this change a user can now
declare the IP-in-IP protocol by using the name 'ipip'

Note: This doesn't fix an underlying problem
with the Neutron protocol number validator, but
serves as a work around for users wanting to use
the IP-in-IP protocol.

Partial-bug: #1716045

Change-Id: I18e5e42b687e12b64f5a9c523a912c8dd1afa9d2