Add new resource for port bindings
As a result of implementing providing Nova port binding information for
live migration [1], a new resource for callbacks is required. This
resource will be used by the multiple port binding service plugin [2]
to notify the core plugin and other subscribers when create, update and
delete operations are performed on a port binding.

[1] https://review.openstack.org/#/c/460021
[2] https://review.openstack.org/#/c/414251

Related-Bug: #1580880

Change-Id: Ibbc7af67cf1282cbf7d9d3a681b9a04a7922c24a