Fix placement calls in placement client
Based on the placement api-ref fix the already implemented placement API
calls in the placement client of neutron-lib.

Fixed/changed parts:
- Add version header to create_resource_provider
- Change create_inventory to update_resource_provider_inventories, as
the http request is a PUT not a POST, and add exception to handle the
case when the rp is not found.
- Change update_inventory to update_resource_provider_inventory, to make
the method name in sync with placement API ref.
- Fix the failing unit tests, and cover the negative case for
update_resource_provider_inventories.

Change-Id: I336c57e734f91795038c5ba9074970d1f5522da2
Related-Bug: #1578989