API attribute processing: add 'default_overrides_none'
This change introduces an 'default_overrides_none' flag.

When enabled, the default value for the attribute will
be used, including if the attribute was explicitly defined
as None.

This is introduced to support the current behavior of the
flowclassifier API extension from networking-sfc
(see Id5aea04553b78dd17fca85de88cf076b42d24140).

Change-Id: I511d7dfd02ceed417d377451ec0265d9704ef261