Remove LABEL from API definitions
LABEL is now useless and plays no role that ALIAS can play

For this reason and since it can be a source of confusion for
people writing API definition files ("what the hell is this used for ?
what am I supposed to put here ?"), this change removes it.

The only places where this remains are https://bit.ly/2Gha5Xw (codesearch).
Cleaning up in these places is handled by the commits
which this change is depending on.

Depends-On: Ida395d94ed5b54ef9c42dc19cf099aab54fd0bb5
Depends-On: I4f0f94ce8d526bf6e1012823409582d46402625b
Depends-On: I96876caaba1e2663eb0451335327cb1276420d4f

Change-Id: I7e2502f8591ad664d20e23e58f33291575f29083